+++ src/main/java/frc/robot/GlobalConstants.java
@@ -1,0 +1,6 @@
+package frc.robot;
+
+public class GlobalConstants {
+    public static final Boolean DEBUG = false;
+    public static final String CTRE_LOG_PATH = "/media/sda1/ctre-logs/";
+}